requirements.txt
aiosmtplib==5.1.0
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==3.2.2
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastapi-mail==1.6.2
fastar==0.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
passlib==1.7.4
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
rich==14.3.1
rich-toolkit==0.17.1
rignore==0.7.6
sentry-sdk==2.50.0
shellingham==1.5.4
SQLAlchemy==2.0.46
sqlmodel==0.0.31
starlette==0.50.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid==1.30
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0

